refactor: Relocate fmx module into Docker/frappe
- Move all files in fmx module to Docker/frappe/fmx
- Update import paths in relocated files
- Add new dependencies (redis, rq) in pyproject.toml
- Define dev dependency group in pyproject.toml
- Remove redundant zshrc alias for "fm" command
- Delete unused `commands/__init__.py` file
- Remove default extensions in frappe_manager init file
