+++ Docker/frappe/fmx/fmx/commands/start.py
@@ -1,24 +1,24 @@
-from typing import Annotated, List, Optional
+from typing import Annotated, Optional, List
 
 import typer
-
-from ..cli import (
+from fmx.display import DisplayManager
+from fmx.command_utils import validate_services, get_process_description
+from fmx.cli import (
     ServiceNameEnumFactory,
     execute_parallel_command,
     get_service_names_for_completion,
 )
-from ..command_utils import get_process_description, validate_services
-from ..display import DisplayManager
-from ..supervisor import stop_service as util_stop_service
+from fmx.supervisor.api import start_service as util_start_service
 
-command_name = "stop"
+command_name = "start"
 
 ServiceNamesEnum = ServiceNameEnumFactory()
+
 
 def command(
     ctx: typer.Context,
     service_names: Annotated[
-        Optional[List[ServiceNamesEnum]],
+        Optional[List[ServiceNamesEnum]], 
         typer.Argument(
             help="Name(s) of the service(s) to target. If omitted, targets ALL running services.",
             autocompletion=get_service_names_for_completion,
@@ -26,10 +26,11 @@ def command(
         )
     ] = None,
     process_name: Annotated[
-        Optional[List[str]],
+        Optional[List[str]], 
         typer.Option(
-            "--process", "-p",
-            help="Target only specific process(es) within the selected service(s). Use multiple times for multiple processes (e.g., -p worker_short -p worker_long).",
+            "--process",
+            "-p",
+            help="Target only specific process(es). If omitted, attempts to start ALL defined processes in the service.",
             show_default=False,
         )
     ] = None,
@@ -40,38 +41,43 @@ def command(
             help="Wait for supervisor start/stop operations to complete before returning.",
         )
     ] = True,
-    wait_workers: Annotated[
-        Optional[bool],
+    verbose: Annotated[
+        bool,
         typer.Option(
-            "--wait-workers/--no-wait-workers",
-            help="Wait for processes identified as workers to stop gracefully (use if default stop times out workers).",
-            show_default=False,
+            "--verbose", "-v",
+            help="Show detailed process identification and skipping messages during start.",
         )
-    ] = None,
+    ] = False,
 ):
-    """Stop services or specific processes."""
+    """Start services or specific processes."""
+
     display: DisplayManager = ctx.obj['display']
 
     all_services = get_service_names_for_completion()
     services_to_target = all_services if not service_names else [s.value for s in service_names]
 
-    valid, target_desc = validate_services(display, services_to_target, all_services, "stop")
+    valid, target_desc = validate_services(display, services_to_target, all_services, "start")
     if not valid:
         return
 
     process_desc = get_process_description(display, process_name)
 
-    wait_desc = "(with wait)" if wait else "(without wait)"
+    if process_name:
+        process_desc = f"specific process(es): {display.highlight(', '.join(process_name))}"
+    else:
+        process_desc = "all defined processes"
 
-    display.print(f"\nAttempting to stop {process_desc} in {target_desc} {wait_desc}...")
+    wait_desc = "(with wait)" if wait else "(without wait)"
+    display.print(f"\nStarting {process_desc} in {target_desc} {wait_desc}...")
+
     execute_parallel_command(
         services_to_target,
-        util_stop_service,
-        action_verb="stopping",
+        util_start_service,
+        action_verb="starting",
         show_progress=True,
         process_name_list=process_name,
         wait=wait,
-        wait_workers=wait_workers
+        verbose=verbose
     )
 
-    display.print("\nStop sequence complete.")
+    display.print("\nStart sequence complete.")
