+++ Docker/frappe/fmx/fmx/commands/status.py
@@ -1,15 +1,15 @@
 from typing import Annotated, Optional, List
 
 import typer
-from ..display import DisplayManager
-from ..command_utils import validate_services
+from fmx.display import DisplayManager
+from fmx.command_utils import validate_services
 
-from ..cli import (
+from fmx.cli import (
     ServiceNameEnumFactory,
     execute_parallel_command,
     get_service_names_for_completion
 )
-from ..supervisor.api import get_service_info as util_get_service_info
+from fmx.supervisor.api import get_service_info as util_get_service_info
 
 command_name = "status"
