+++ Docker/frappe/fmx/fmx/supervisor/stop_helpers.py
@@ -1,14 +1,14 @@
 import logging
 import time
-from typing import Optional, List, Any, Dict
+from typing import Optional, Any, Dict
 from xmlrpc.client import Fault
 
-from ..display import display
+from fmx.display import display
 
 logger = logging.getLogger(__name__)
 
-from .constants import STOPPED_STATES, is_worker_process
-from .fault_handler import _raise_exception_from_fault
+from fmx.supervisor.constants import STOPPED_STATES, is_worker_process
+from fmx.supervisor.fault_handler import _raise_exception_from_fault
 
 def _wait_for_process_stop(supervisor_api, process_name: str, timeout: int) -> bool:
     """Wait for a single process to reach a stopped state."""
