+++ Docker/frappe/fmx/pyproject.toml
@@ -33,6 +33,8 @@ dependencies = [
     "typer>=0.16.0",
     "rich>=13.0.0,<15.0.0",
     "supervisor>=4.2.0,<5.0.0",
+    "redis>=6.1.1",
+    "rq>=2.3.3",
 ]
 
 [project.urls]
@@ -51,6 +53,11 @@ exclude = ["tests*"]
 
 [tool.setuptools.package-data]
 "fmx" = ["py.typed"]
+
+[dependency-groups]
+dev = [
+    "nuitka>=2.7.12",
+]
 
 [tool.setuptools]
 zip-safe = false
