--- fmx/fmx/commands/__init__.py
@@ -1,1 +1,0 @@
-# This file makes Python treat the directory 'commands' as a package.
